requirements.txt
h5py~=3.12.1
polars~=1.18.0
omegaconf~=2.3.0
timm~=1.0.11
einops~=0.7.0
seaborn~=0.13.2
wandb~=0.19.1
matplotlib~=3.8.3
hydra-core~=1.3.2
torchmetrics~=1.3.2
tqdm~=4.66.4
optuna~=4.2.1